\name{StatComp21007-package}
\alias{StatComp21007-package}
\alias{StatComp21007}
\docType{package}
\title{
  Final homework of statistical computations
}
\description{
  It includes the homework of the whole semester and two functions of my research field.
}
\author{
StatComp21007
Maintainer: StatComp21007 <WWWang0903@126.com>
}
\keyword{ package }
